refactor(dashboard): remove ValidateSession
